Split filter code into files
